Login/Registration Functionality (No Database connection)
User registration implemented with methods to create a password salt and hash at the time of registration IF both password fields match.

User login function checks if the username exists in the database; if it does, then take the given password and hash it with that user's salt. Validate the user if hash is correct.

Notes:
1.  connection string in Server/UserDbContext.cs needs to be changed once database connection is established.
2. Testing is needed once database is connected to ensure methods definitely work.
3. Login token implementation needed.